transposition table

+++ Result.java
@@ -2,6 +2,6 @@ import com.github.bhlangonijr.chesslib.move.Move;
 
 class Result
 {
-	int score;
+	short score;
 	Move m;
 }
